Removed websteps.rb
I found out that websteps.rb is bad, so I removed it.
