+++ overlord/Gemfile
@@ -13,3 +13,4 @@ gem 'rack-test'
 gem 'guard'
 gem 'guard-cucumber'
 gem 'foreman'
+gem 'selenium-webdriver'
